rsockets: Allow user to specify the QP size
Add a setsockopt that allows the user to specify the desired
size of the underlying QP.  The provided size is used as the
maximum size when creating the QP.  The actual size of the QP
is the smaller of the user provided maximum and the maximum
size supported by the underlying hardware.

A user may retrieve the actual size of the QP through the
getsockopt call.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>