rsocket: Spin before blocking on an rsocket
The latency cost of blocking is significant compared to round
trip ping-pong time.  Spin briefly on rsockets before calling
into the kernel and blocking.

We calculate the time to delay blocking by timing the round
trip ping-pong time on a QP configured for loopback traffic.  The
delay time is a multiple of this time to include wire transfer
time, plus minimal response time on the remote side.

As a completely unintentional side effect, this just happens to
improve application performance in benchmarks, like netpipe.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>