rsocket: Reserve bits in rsocket message
rsockets includes the length of a data transfer in the
immediate data.  Currently 29 bits are reserved for the
length, but the maximum transfer is capped at 64K.
For future expansion, limit the length field
to 17 bits and reserve the other 12.

The interpretation of the 17 bits is modified as follows.
The lower 16 bits indicate a length, with the upper bit
indicating a scaling factor.  If the upper bit is
0, the scaling factor is 1.  The lower 16 bits indicate the
transfer size in bytes.  If the upper bit is 1, then a
scaling factor of 64K is applied.  The total transfer size is
calculated as (value carried in bits 0-15 + 1) * 64 KB.

This is backwards compatible with the current implementation.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>