acm: Define needed ACM protocol messages
The librdmacm needs message definitions used to communicate
with the ibacm.  It currently pulls these from infiniband/acm.h,
which is installed by ibacm.  This creates an install order
dependency on ibacm.  However, work on the scalable SA has
the ibacm using the librdmacm (via rsockets) for communication
between the different SSA components.

To resolve this issue, have the librdmacm define the message
structures that it needs to communicate with ibacm.  The
librdmacm already defines some ACM messages through configuration
checks.  We just expand that capability, which isolates the librdmacm
package from the ibacm package.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>